+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **The demo collection travels to Bangkok**: Shophouse Rooftop Studios, over Charoen Krung, replaces the Tel Aviv rooftop-studios project
+
 ## [0.12.0] - 2026-08-20
 
 ### Added